# Process files in parallel