@q3cpma 2023-07-09
Pure curiosity, is there some kind of thread pool sharing between files?