@q3cpma 2023-07-09
On Sun Jul 9, 2023 at 10:13 PM CEST, andrews05 wrote:
> @q3cpma Yes, oxipng uses Rayon with a single thread pool which is used for all parallel tasks, including reduction evaluations and compression trials. Since this was already a part of oxipng, the change required here was essentially just replacing the files iterator with a parallel iterator. Rayon handles the workload automatically - see here for more info: https://github.com/rayon-rs/rayon/blob/master/FAQ.md

Cool! I kinda expected this, knowing the Rust ecosystem, but still worth asking.