@TPS 2023-07-11
I'm sorry, just saw this & read tne `rayon` ReadMe, which jogged a memory:

There it mentions having to do special adaptations for `wasm` targets via [`wasm-bindgen-rayon`](https://github.com/GoogleChromeLabs/wasm-bindgen-rayon).  I know a while ago some special work was done in `oxipng` to get that working.  Does it need to be adjusted in light of this & related?