@shssoichiro 2023-07-11
I'm not sure entirely about wasm-bindgen-rayon, the last of my knowledge is that wasm does not have proper threading support and so wasm users will need to disable the "parallel" feature in oxipng. It's possible that may have changed and there's a way to support threading on wasm now, someone more specialized in wasm may know.